UI refactoring started
