+++ src/apdu_handlers.c
@@ -1,5 +1,8 @@
 #include "apdu_handlers.h"
 
+// commandContext global;
+// ux_state_t ux;
+
 handler_fn_t *lookupHandler(uint8_t ins) {
     switch (ins) {
         case INS_GET_VERSION:
@@ -20,15 +23,15 @@ void handleGetPublicKey(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t da
     cx_ecfp_public_key_t publicKey;
     cx_ecfp_private_key_t privateKey;
 
-    unsigned int bip32_path[BIP44_PATH_LEN];
+    unsigned int bip44_path[BIP44_PATH_LEN];
     uint32_t i;
-    for (i = 0; i < BIP32_PATH_LEN; i++) {
-        bip32_path[i] = (dataBuffer[0] << 24) | (dataBuffer[1] << 16) | (dataBuffer[2] << 8) |
+    for (i = 0; i < BIP44_PATH_LEN; i++) {
+        bip44_path[i] = (dataBuffer[0] << 24) | (dataBuffer[1] << 16) | (dataBuffer[2] << 8) |
                         (dataBuffer[3]);
         dataBuffer += 4;
     }
     unsigned char privateKeyData[32];
-    os_perso_derive_node_bip32(CX_CURVE_256K1, bip32_path, BIP32_PATH_LEN, privateKeyData,
+    os_perso_derive_node_bip32(CX_CURVE_256K1, bip44_path, BIP44_PATH_LEN, privateKeyData,
                                NULL);
     cx_ecdsa_init_private_key(CX_CURVE_256K1, privateKeyData, 32, &privateKey);
 
@@ -40,9 +43,8 @@ void handleGetPublicKey(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t da
     os_memmove(G_io_apdu_buffer, publicKey.W, 65);
     *tx = 65;
 
-    display_address(publicKey.W);
-    refresh_public_key_display();
-    screen_printf("Address: %s\n", address);
+    display_address(publicKey.W, global.getPublicKeyContext.address);
+    screen_printf("Address: %s\n", global.getPublicKeyContext.address);
 
     THROW(INS_RET_SUCCESS);
 }
@@ -61,10 +63,4 @@ void handleGetSignedPublicKey(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint1
 
 void handleGetAddress(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags,
                       volatile unsigned int *tx) {
-}
-
-static void refresh_public_key_display(void) {
-    if ((uiState == UI_PUBLIC_KEY_1) || (uiState == UI_PUBLIC_KEY_2)) {
-        publicKeyNeedsRefresh = 1;
-    }
 }
